refactor: 주석 추가 및 코드 리팩토링
